Add a Postgres GIS HA Container and associated example
In keeping with the Postgres HA update, the Postgres GIS container
and example need to be created using this new image as the base.

This update creates the new container's Centos and RHEL7 Dockerfiles,
for Postgres version 11, which use the respective Centos and RHEL7
Postgres HA containers as their base. It updates the setup.sql to match
the behavior of the new container and creates a separate example.
Finally, it updates the Makefile to create a new Make target.