Remove the Use of curl for Interacting With Patroni
This commit fully removes the use of 'curl' for interacting with
Patroni inside of the 'crunchy-postgres-ha container.  Specifically,
interactions with the Patroni REST API, such as when monitoring
cluster status during initialization or when reloading local
configuration following a config file change (among other examples),
are now done via the 'patronictl' client utility.  This ensures that
the various interactions with the Patroni REST API are able to work
seamlessly across different Patroni server configurations since the
'patronictl' client is able to automatically configure itself
according to any applicable REST API settings and therefore
successfully interact with the Patroni REST API.